requirements.txt
torch>=1.12
torchvision>=0.13
scikit-learn>=1.0
pandas>=1.3
numpy>=1.21
matplotlib>=3.4
opencv-python>=4.5
tqdm>=4.62
Pillow>=8.3
openpyxl>=3.0  # for reading .xlsx metadata
